+++ database/migrations/2019_05_06_072206_create_codes_table.php
@@ -10,7 +10,7 @@ class CreateCodesTable extends Migration
     {
         Schema::create('codes', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('code')->unique();
+            $table->unsignedBigInteger('code')->unique();
         });
     }
